+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Use proxy excludes from global config #115
+
 ## [3.11.5-1] 2026-02-25
 ### Changed
 - Upgrade SCM-Manager to version 3.11.5 ([Changelog](https://github.com/scm-manager/scm-manager/blob/3.11.5/CHANGELOG.md))